refactor: move utility functions from useForm to utils
