+++ .changeset/famous-rice-know.md
@@ -1,0 +1,5 @@
+---
+"react-cool-form": patch
+---
+
+refactor: move utility functions from useForm to utils
