+++ src/useForm.ts
@@ -1,4 +1,4 @@
-import { useCallback, useEffect, useLayoutEffect, useRef } from "react";
+import { useCallback, useEffect, useRef } from "react";
 
 import {
   Config,
@@ -31,6 +31,7 @@ import {
   isArray,
   isCheckboxField,
   isEmptyObject,
+  isFieldElement,
   isFileField,
   isFileList,
   isFunction,
@@ -40,32 +41,12 @@ import {
   isRadioField,
   isRangeField,
   isUndefined,
+  runWithLowPriority,
   set,
   unset,
+  useUniversalLayoutEffect,
   warn,
 } from "./utils";
-
-const useUniversalLayoutEffect =
-  typeof window === "undefined" ? useEffect : useLayoutEffect;
-
-const runWithLowPriority = (callback: (args: any) => any) =>
-  (
-    window.requestIdleCallback ||
-    ((callback) => {
-      const start = Date.now();
-      return setTimeout(
-        () =>
-          callback({
-            didTimeout: false,
-            timeRemaining: () => Math.max(0, 50 - (Date.now() - start)),
-          }),
-        1
-      );
-    })
-  )(callback, { timeout: 2000 });
-
-const isFieldElement = ({ tagName }: HTMLElement) =>
-  /INPUT|TEXTAREA|SELECT/.test(tagName);
 
 export default <V extends FormValues = FormValues>({
   defaultValues,
